SemiWorking
Using the test file provided, a ";" needs to be added to line 5, and a "." added to line 19. If those two things are done the tree is built correctly. The "==>" is searched for if the when is inside the Loop. Not sure how to handle the when statement outside the loop. Page is only letting me upload < 100 files, so i'm just uploading wci folder.